requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-filter>=23.0
django-extensions>=3.2.0
Pillow>=10.0.0
python-decouple>=3.8
psycopg2-binary>=2.9.0
celery>=5.3.0
redis>=4.5.0
django-celery-beat>=2.5.0
django-celery-results>=2.5.0
requests>=2.31.0
plotly>=5.15.0
pandas>=2.0.0
numpy>=1.24.0
django-model-utils>=4.2.0
